[refactor]: user관련 dto는 user 패키지에서 관리
