+++ src/main/java/com/dku/springstudy/controller/UserController.java
@@ -2,6 +2,7 @@ package com.dku.springstudy.controller;
 
 import com.dku.springstudy.config.security.jwt.JwtProvider;
 import com.dku.springstudy.dto.*;
+import com.dku.springstudy.dto.user.*;
 import com.dku.springstudy.model.User;
 import com.dku.springstudy.repository.UserRepository;
 import com.dku.springstudy.service.S3Service;
