+++ src/main/java/com/dku/springstudy/dto/user/UserInformationChangeRequestDTO.java
@@ -1,4 +1,4 @@
-package com.dku.springstudy.dto;
+package com.dku.springstudy.dto.user;
 
 import lombok.Builder;
 import lombok.Data;
